Spline:  * added spline interploations also for md values  * added boolean property "refined"  * spline functions in separate file
