+++ react/src/lib/components/DeckGLMap/layers/wells/wellsLayer.ts
@@ -9,7 +9,7 @@ import { color } from "d3-color";
 import { Feature } from "geojson";
 import { LayerPickInfo, PropertyDataType } from "../utils/layerTools";
 import { patchLayerProps } from "../utils/layerTools";
-import { Position2D, Position3D } from "@deck.gl/core/utils/positions";
+import { splineRefine } from "./utils/spline";
 
 export interface WellsLayerProps<D> extends CompositeLayerProps<D> {
     pointRadiusScale: number;
@@ -22,6 +22,7 @@ export interface WellsLayerProps<D> extends CompositeLayerProps<D> {
     logrunName: string;
     logRadius: number;
     logCurves: boolean;
+    refined: boolean;
 }
 
 const defaultProps = {
@@ -69,9 +70,13 @@ export default class WellsLayer extends CompositeLayer<
     }
 
     renderLayers(): (GeoJsonLayer<Feature> | PathLayer<LogCurveDataType>)[] {
+
+        // Input 3D wellpaths will be projected to z = 0 plane (setting z values to zero).
+        // Also if "refined" is specified it will interpolate survey points and refine path using
+        // spline interpolation.
+        const refined = this.props.refined;
         //const now = Date.now();
-        const refine = true;
-        const data = splineRefine(this.props.data, refine);
+        const data = splineRefine(this.props.data, refined);
         //console.log("time elapsed:", Date.now() - now);
 
         const outline = new GeoJsonLayer<Feature>(
@@ -352,160 +357,4 @@ function getLogProperty(
 
     if (log_value) return { name: prop_name, value: log_value };
     else return null;
-}
-
-function CatmullRom(
-    P0: Position3D,
-    P1: Position3D,
-    P2: Position3D,
-    P3: Position3D,
-    t: number
-): Position3D {
-    const alpha = 0.5;
-    const tt = t * t;
-    const ttt = t * t * t;
-
-    // disable eslint for some lines due to readability.
-    const dist_p0_p1 = Math.sqrt((P1[0]-P0[0])*(P1[0]-P0[0]) + (P1[1]-P0[1])*(P1[1]-P0[1]) + (P1[2]-P0[2])*(P1[2]-P0[2]) ); // eslint-disable-line
-    const dist_p1_p2 = Math.sqrt((P1[0]-P2[0])*(P1[0]-P2[0]) + (P1[1]-P2[1])*(P1[1]-P2[1]) + (P1[2]-P2[2])*(P1[2]-P2[2]) ); // eslint-disable-line
-    const dist_p2_p3 = Math.sqrt((P3[0]-P2[0])*(P3[0]-P2[0]) + (P3[1]-P2[1])*(P3[1]-P2[1]) + (P3[2]-P2[2])*(P3[2]-P2[2]) ); // eslint-disable-line
-
-    const t01 = Math.pow(dist_p0_p1, alpha);
-    const t12 = Math.pow(dist_p1_p2, alpha);
-    const t23 = Math.pow(dist_p2_p3, alpha);
-
-    const m1_x = (P2[0] - P1[0] + t12 * ((P1[0] - P0[0]) / t01 - (P2[0] - P0[0]) / (t01 + t12))); // eslint-disable-line
-    const m1_y = (P2[1] - P1[1] + t12 * ((P1[1] - P0[1]) / t01 - (P2[1] - P0[1]) / (t01 + t12))); // eslint-disable-line
-    const m1_z = (P2[2] - P1[2] + t12 * ((P1[2] - P0[2]) / t01 - (P2[2] - P0[2]) / (t01 + t12))); // eslint-disable-line
-
-    const m2_x = (P2[0] - P1[0] + t12 * ((P3[0] - P2[0]) / t23 - (P3[0] - P1[0]) / (t12 + t23))); // eslint-disable-line
-    const m2_y = (P2[1] - P1[1] + t12 * ((P3[1] - P2[1]) / t23 - (P3[1] - P1[1]) / (t12 + t23))); // eslint-disable-line
-    const m2_z = (P2[2] - P1[2] + t12 * ((P3[2] - P2[2]) / t23 - (P3[2] - P1[2]) / (t12 + t23))); // eslint-disable-line
-
-    const a_x = 2 * (P1[0] - P2[0]) + m1_x + m2_x;
-    const a_y = 2 * (P1[1] - P2[1]) + m1_y + m2_y;
-    const a_z = 2 * (P1[2] - P2[2]) + m1_z + m2_z;
-
-    const b_x = -3 * (P1[0] - P2[0]) - m1_x - m1_x - m2_x;
-    const b_y = -3 * (P1[1] - P2[1]) - m1_y - m1_y - m2_y;
-    const b_z = -3 * (P1[2] - P2[2]) - m1_z - m1_z - m2_z;
-
-    const c_x = m1_x;
-    const c_y = m1_y;
-    const c_z = m1_z;
-
-    const d_x = P1[0];
-    const d_y = P1[1];
-    const d_z = P1[2];
-
-    const x = a_x * ttt + b_x * tt + c_x * t + d_x;
-    const y = a_y * ttt + b_y * tt + c_y * t + d_y;
-    const z = a_z * ttt + b_z * tt + c_z * t + d_z;
-
-    return [x, y, z] as Position3D;
-}
-
-function splineRefine(data, refine: boolean) {
-    const ts = refine ? [0.2, 0.4, 0.6, 0.8] : [];
-
-    if (data["features"] === undefined) {
-        return;
-    }
-
-    const no_wells = data["features"].length;
-    for (let well_no = 0; well_no < no_wells; well_no++) {
-        const mds = data["features"][well_no]["properties"]["md"];
-
-        const coords = data["features"][well_no]["geometry"]["geometries"][1]["coordinates"]; // eslint-disable-line
-
-        const n = coords.length;
-        if (n < 3) {
-            continue;
-        }
-
-        // Point before first.
-        const x0 = coords[0][0] - coords[1][0] + coords[0][0];
-        const y0 = coords[0][1] - coords[1][1] + coords[0][1];
-        const z0 = coords[0][2] - coords[1][2] + coords[0][2];
-        const P_first: Position3D = [x0, y0, z0];
-
-        // Point after last.
-        const xn = coords[n - 1][0] - coords[n - 2][0] + coords[n - 1][0];
-        const yn = coords[n - 1][1] - coords[n - 2][1] + coords[n - 1][1];
-        const zn = coords[n - 1][2] - coords[n - 2][2] + coords[n - 1][2];
-        const P_n: Position3D = [xn, yn, zn];
-
-        //const md_first = 0.25 * (mds[0][0] - mds[0][1]) + mds[0][0];
-        //const md_n = 0.25 * (mds[0][n - 1] - mds[0][n - 2]) + mds[0][n - 1];
-
-        const newCoordinates: [Position3D?] = [];
-        const newMds: number[][] = [];
-        newMds.push([]);
-
-        for (let i = 0; i < n - 2; i += 1) {
-            let P0: Position3D, P1: Position3D, P2: Position3D, P3: Position3D;
-            //let md0: number;
-            let md1: number;
-            //let md2: number;
-            //let md3: number;
-
-            if (i === 0) {
-                P0 = P_first;
-                P1 = coords[i + 0];
-                P2 = coords[i + 1];
-                P3 = coords[i + 2];
-
-                //md0 = md_first;
-                md1 = mds[0][i + 0];
-                //md2 = mds[0][i + 1];
-                //md3 = mds[0][i + 2];
-            } else if (i === n - 3) {
-                P0 = coords[i + 0];
-                P1 = coords[i + 1];
-                P2 = coords[i + 2];
-                P3 = P_n;
-
-                //md0 = mds[0][i + 0];
-                md1 = mds[0][i + 1];
-                //md2 = mds[0][i + 2];
-                //md3 = md_n;
-            } else {
-                P0 = coords[i + 0];
-                P1 = coords[i + 1];
-                P2 = coords[i + 2];
-                P3 = coords[i + 3];
-
-                //md0 = mds[0][i + 0];
-                md1 = mds[0][i + 1];
-                //md2 = mds[0][i + 2];
-                //md3 = mds[0][i + 3];
-            }
-
-            newCoordinates.push(P1);
-            newMds[0].push(md1);
-
-            for (let t_i = 0; t_i < ts.length; t_i += 1) {
-                const t = ts[t_i];
-                const [x, y, z] = CatmullRom(P0, P1, P2, P3, t);
-                const md = mds[0][i];
-
-                newCoordinates.push([x, y, z] as Position3D);
-                newMds[0].push(md);
-            }
-        }
-
-        newCoordinates.push(coords[n - 1]);
-        newMds[0].push(mds[n - 1]);
-
-        // Convert well path to 2D.
-        const coords2D: Position2D[] = newCoordinates.map((e: Position3D) => {
-            return [e[0], e[1]] as Position2D;
-        });
-
-        data["features"][well_no]["geometry"]["geometries"][1]["coordinates"] =
-            coords2D;
-        data["features"][well_no]["properties"]["md"] = newMds;
-    }
-
-    return data;
 }
